+++ WebContent/WEB-INF/templates/cart.html
@@ -49,7 +49,7 @@
 
 		<div class="input-field white-text" id="wc24-ship">
 			<i class="material-icons prefix">contact_mail</i>
-			<input id="recipient" type="text" class="validate white-text" data-length="16">
+			<input id="recipient" name="recipient" type="text" class="validate white-text" data-length="16">
 			<label for="recipient" class="">Wii Friend Code</label>
 			This will send the content to your Wii through RiiConnect24. <br/> Make sure you're friends with us and have DIY Showcase installed!
 		</div>
