+++ pom.xml
@@ -2,7 +2,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>DoujinSoft</groupId>
   <artifactId>DoujinSoft</artifactId>
-  <version>1.3.0</version>
+  <version>2.0.0</version>
   <packaging>war</packaging>
   <name>DoujinSoft Store</name>
   <build>
@@ -51,11 +51,6 @@
 	    <version>1.0</version>
 	</dependency>
 	<dependency>
-	    <groupId>lzss</groupId>
-	    <artifactId>lzss-native</artifactId>
-	    <version>1.0</version>
-	</dependency>  
-	<dependency>
         <groupId>javax.servlet</groupId>
         <artifactId>javax.servlet-api</artifactId>
         <version>3.1.0</version>
@@ -81,6 +76,16 @@
         <artifactId>httpclient</artifactId>
         <version>4.5.9</version>
     </dependency>
+      <dependency>
+          <groupId>org.apache.httpcomponents</groupId>
+          <artifactId>httpmime</artifactId>
+          <version>4.5.9</version>
+      </dependency>
+      <dependency>
+        <groupId>net.java.dev.jna</groupId>
+        <artifactId>jna</artifactId>
+        <version>4.3.0</version>
+      </dependency>
   </dependencies>
   
   <url>https://github.com/Difegue/DoujinSoft</url>
